+++ CHANGELOG.md
@@ -8,11 +8,12 @@ If you like this project and find it useful, please consider giving it a star on
 
 ### Added
 - [device]: Added addRequiredClusterServers and addOptionalClusterServers methods.
+- [frontend]: Added separated settings for the two logging systems (Matterbridge and Matter.js).
 
 ### Changed
 
 - [device]: Refactor contructor and loadInstance to accept DeviceTypeDefinition | AtLeastOne<DeviceTypeDefinition>.
-- [frontend]: Update to 1.2.0 (optimization for mobile)
+- [frontend]: Update to 1.2.0 (initial optimization for mobile)
 
 ## [1.3.10] - 2024-07-05